refactor(bkn-agent): rename service agent-runtime -> bkn-agent
Service naming decision: the platform-internal agent runtime (Epic #202)
is named bkn-agent. Mechanical rename, no behavior change:

- infra/agent-runtime/ -> infra/bkn-agent/ (chart name, k8s service,
  image repo, env prefix AGENT_RUNTIME_* -> BKN_AGENT_*, error code
  prefix AgentRuntime.* -> BknAgent.*, API prefix /api/bkn-agent/v1)
- docs/api/agent-runtime.yaml -> docs/api/bkn-agent.yaml (re-exported
  from app, contract test green)
- docs/design/agent-runtime/ -> docs/design/bkn-agent/
- migrations/agent-runtime/ -> migrations/bkn-agent/ (agent_ table
  prefix unchanged per design)

trace-ai docs' "Agent Runtime (Dolphin)" refers to the historical
product and is intentionally untouched.

Tests: 18 passed (pytest, py3.12 venv).
