--- infra/agent-runtime/charts/Chart.yaml
@@ -1,5 +1,0 @@
-apiVersion: v1
-appVersion: "1.0"
-description: Platform-internal agent runtime (Epic #202)
-name: agent-runtime
-version: 0.1.0
