--- infra/agent-runtime/charts/values.yaml
@@ -1,58 +1,0 @@
-replicaCount: 1
-
-namespace: openbkn
-
-image:
-  registry: ghcr.io/openbkn-ai
-  repository: agent-runtime
-  tag: "__VERSION__"
-  pullPolicy: IfNotPresent
-  name: agent-runtime
-
-service:
-  type: ClusterIP
-  ports:
-    port: 30800
-    targetPort: 30800
-
-# 仅面向平台内部（Epic #202 硬约束）：本 chart 不含 ingress 模板，
-# 网关不得为 agent-runtime 开终端用户入口。产品面需要 agent 能力时
-# 由其后端以服务身份代理调用。
-
-depServices:
-  rds:
-    host: "rds-proton-rds-mariadb-cluster"
-    port: "3306"
-    user: "root"
-    password: "password"
-    database: "openbkn"
-
-runtime:
-  mfModelApiPrivateBase: "http://mf-model-api:9898/api/private/mf-model-api/v1"
-  agentRetrievalMcpUrl: "http://agent-retrieval:30779/api/agent-retrieval/v1/mcp"
-  capabilitiesLabBase: "http://capabilities-lab:8080/api/capabilities-lab/v1"
-  # published agent 注册进算子工厂 toolbox（#212）
-  operatorIntegrationBase: "http://agent-operator-integration:9000"
-  toolboxSync: "true"
-  # toolbox 工具回调本服务的地址（box_svc_url）
-  selfBaseUrl: "http://agent-runtime:30800"
-  defaultModel: ""
-  checkpointerBackend: "mysql"
-  # 建表统一归 migrations/agent-runtime/（core-data-migrator）。
-  # checkpointer 三张表（langgraph-checkpoint-mysql 固定表名）当前由首启
-  # 一次性开启此开关建出，后续固化进迁移后关回 false。
-  checkpointerAllowRuntimeDdl: "false"
-
-observability:
-  otelEnabled: "true"
-  otlpEndpoint: "http://otelcol-contrib:4318"
-
-nodeSelector: {}
-
-resources:
-  requests:
-    cpu: 250m
-    memory: 512Mi
-  limits:
-    cpu: 2
-    memory: 2Gi
