+++ infra/bkn-agent/charts/Chart.yaml
@@ -1,0 +1,5 @@
+apiVersion: v1
+appVersion: "1.0"
+description: Platform-internal agent runtime "bkn-agent" (Epic #202)
+name: bkn-agent
+version: 0.1.0
