+++ infra/bkn-agent/charts/templates/deployment.yaml
@@ -1,0 +1,93 @@
+apiVersion: apps/v1
+kind: Deployment
+metadata:
+  name: {{ .Values.image.name }}
+  namespace: {{ .Values.namespace }}
+spec:
+  replicas: {{ .Values.replicaCount }}
+  selector:
+    matchLabels:
+      app: {{ .Values.image.name }}
+  strategy:
+    rollingUpdate:
+      maxSurge: 0
+      maxUnavailable: 25%
+    type: RollingUpdate
+  template:
+    metadata:
+      labels:
+        app: {{ .Values.image.name }}
+    spec:
+      dnsConfig:
+        options:
+          - name: ndots
+            value: "1"
+      {{- with .Values.nodeSelector }}
+      nodeSelector:
+        {{- toYaml . | nindent 8 }}
+      {{- end }}
+      containers:
+        - name: {{ .Chart.Name }}
+          resources:
+            {{- toYaml .Values.resources | nindent 12 }}
+          image: "{{- if .Values.image.registry }}{{ .Values.image.registry }}/{{ end }}{{ .Values.image.repository }}:{{ .Values.image.tag }}"
+          imagePullPolicy: {{ .Values.image.pullPolicy }}
+          ports:
+            - name: http
+              containerPort: {{ .Values.service.ports.targetPort }}
+          livenessProbe:
+            httpGet:
+              path: /api/v1/health
+              port: {{ .Values.service.ports.targetPort }}
+            initialDelaySeconds: 30
+            periodSeconds: 60
+            timeoutSeconds: 10
+          startupProbe:
+            httpGet:
+              path: /api/v1/health
+              port: {{ .Values.service.ports.targetPort }}
+            initialDelaySeconds: 5
+            periodSeconds: 10
+            failureThreshold: 30
+          env:
+            - name: BKN_AGENT_PORT
+              value: {{ .Values.service.ports.targetPort | quote }}
+            - name: RDSHOST
+              value: {{ .Values.depServices.rds.host | quote }}
+            - name: RDSPORT
+              value: {{ .Values.depServices.rds.port | quote }}
+            - name: RDSDBNAME
+              value: {{ .Values.depServices.rds.database | quote }}
+            - name: RDSUSER
+              value: {{ .Values.depServices.rds.user | quote }}
+            - name: RDSPASS
+              value: {{ .Values.depServices.rds.password | quote }}
+            - name: MF_MODEL_API_PRIVATE_BASE
+              value: {{ .Values.runtime.mfModelApiPrivateBase | quote }}
+            - name: AGENT_RETRIEVAL_MCP_URL
+              value: {{ .Values.runtime.agentRetrievalMcpUrl | quote }}
+            - name: CAPABILITIES_LAB_BASE
+              value: {{ .Values.runtime.capabilitiesLabBase | quote }}
+            - name: OPERATOR_INTEGRATION_BASE
+              value: {{ .Values.runtime.operatorIntegrationBase | quote }}
+            - name: BKN_AGENT_TOOLBOX_SYNC
+              value: {{ .Values.runtime.toolboxSync | quote }}
+            - name: BKN_AGENT_SELF_BASE_URL
+              value: {{ .Values.runtime.selfBaseUrl | quote }}
+            - name: BKN_AGENT_DEFAULT_MODEL
+              value: {{ .Values.runtime.defaultModel | quote }}
+            - name: CHECKPOINTER_BACKEND
+              value: {{ .Values.runtime.checkpointerBackend | quote }}
+            - name: CHECKPOINTER_ALLOW_RUNTIME_DDL
+              value: {{ .Values.runtime.checkpointerAllowRuntimeDdl | quote }}
+            - name: OTEL_ENABLED
+              value: {{ .Values.observability.otelEnabled | quote }}
+            - name: OTEL_EXPORTER_OTLP_ENDPOINT
+              value: {{ .Values.observability.otlpEndpoint | quote }}
+          volumeMounts:
+            - name: host-time
+              mountPath: /etc/localtime
+      volumes:
+        - name: host-time
+          hostPath:
+            path: /etc/localtime
